remove useless code
