+++ tests/drainer_info.go
@@ -14,12 +14,15 @@
 package tests
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"io/ioutil"
 	"os/exec"
 	"strings"
+	"text/template"
 
+	"github.com/pingcap/tidb-operator/tests/slack"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/kubernetes/test/e2e/framework/log"
@@ -37,6 +40,34 @@ const (
 	// DbTypeKafka DbType = "kafka"
 	// DbTypeFlash DbType = "flash"
 )
+
+var sqlDrainerConfigTemp string = drainerConfigCommon + `
+  db-type = "{{ .DbType }}"
+  [syncer.to]
+  host = "{{ .Host }}"
+  user = "{{ .User }}"
+  password = "{{ .Password }}"
+  port = {{ .Port }}
+`
+
+var fileDrainerConfigTemp string = drainerConfigCommon + `
+  db-type = "file"
+  [syncer.to]
+  dir = "/data/pb"
+`
+
+var drainerConfigCommon string = `
+initialCommitTs: "{{ .InitialCommitTs }}"
+config: |
+  detect-interval = 10
+  compressor = ""
+  [syncer]
+  worker-count = 16
+  disable-dispatch = false
+  ignore-schemas = "INFORMATION_SCHEMA,PERFORMANCE_SCHEMA,mysql"
+  safe-mode = false
+  txn-batch = 20
+`
 
 type DrainerSourceConfig struct {
 	Namespace      string
@@ -159,4 +190,34 @@ func (d *DrainerConfig) DrainerHelmString(m map[string]string, source *DrainerSo
 		arr = append(arr, fmt.Sprintf("%s=%s", k, v))
 	}
 	return strings.Join(arr, ",")
+}
+
+func GetDrainerSubValuesOrDie(info *DrainerConfig) string {
+	if info == nil {
+		slack.NotifyAndPanic(fmt.Errorf("Cannot get drainer sub values, the drainer config is nil"))
+	}
+	buff := new(bytes.Buffer)
+	switch info.DbType {
+	case DbTypeFile:
+		temp, err := template.New("file-drainer").Parse(fileDrainerConfigTemp)
+		if err != nil {
+			slack.NotifyAndPanic(err)
+		}
+		if err := temp.Execute(buff, &info); err != nil {
+			slack.NotifyAndPanic(err)
+		}
+	case DbTypeTiDB:
+		fallthrough
+	case DbTypeMySQL:
+		temp, err := template.New("sql-drainer").Parse(sqlDrainerConfigTemp)
+		if err != nil {
+			slack.NotifyAndPanic(err)
+		}
+		if err := temp.Execute(buff, &info); err != nil {
+			slack.NotifyAndPanic(err)
+		}
+	default:
+		slack.NotifyAndPanic(fmt.Errorf("db-type %s has not been suppored yet", info.DbType))
+	}
+	return buff.String()
 }
